Rework how options are parsed
It used to be that each "pair" was given as a vim syntax command
(e.g. "start=/(/ end=/)/") which was then parsed, to find if it defined
any containedin or contains options, since these need to be treated
specially.  You could also do ['(', ')'] for the same effect, but this
wasn't done anywhere in the default options.  This previous List could
also define an operator in the middle (e.g. ['{', ':', '}']) but that
was conceptually broken, since that operator would not be highlighted
just inside of that pair but everywhere.

This change changes all that.  You can now either continue doing what
you were doing ("start=... end=..."), or use a List, like before, though
without the ability to define operators as the middle element of the
List.  You can also use a Dictionary like so:
    {"start": "/(", "end": "/)", "contains": "...", "fold": 1}
The keys are options of the syntax command.  For boolean options (like
the "fold" above), you set them to 1.  You could also set them to 0,
it's the same as not defining them at all.

Internally, all of the three option types are turned into a Dictionary,
from which the containedin and contains options are extracted (they are
still special), and everything else too.