Rework how the plugin and the options are loaded
No more rainbow#hook, rainbow#load.  What used to be #hook is now #load,
and it's the way to activate the plugin.  It deals with the merging of
the configurations and then the setting up the syntax groups and the
highlighting.  Toggling, therefore, first tears everything down fully,
and then loads it all up again.

Also, remove b:rainbow_visible.  I don't really understand what the need
for it was.  b:loaded_rainbow (the new name for b:rainbow_loaded, to be
consistent with the g: variant), did almost the same job, and some more.

rainbow#show and rainbow#hide are renamed, though I kinda want to make
them script-local, and let them take their needed info through
arguments, rather that buffer-local variables.  That would also mean
that they would only be called by #load and #unload, which would
simplify stuff a bit.